backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
astunparse==1.6.3
attrs==26.1.0
biopython==1.87
biotite==1.6.0
biotraj==1.2.2
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
decorator==5.2.1
einops==0.8.2
executing==2.2.1
fair-esm==2.0.0
fastapi==0.135.3
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httptools==0.7.1
huggingface_hub==0.36.2
idna==3.11
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
msgpack-numpy==0.4.8
namex==0.1.0
networkx==3.6.1
numpy
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
prompt_toolkit==3.0.52
protobuf==7.34.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.11.0
torchaudio==2.11.0
torchtext==0.18.0
torchvision==0.26.0
tqdm==4.67.3
traitlets==5.14.3
transformers==4.48.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wrapt==2.1.2
zstd==1.5.7.3
